HunterCelery/requirements.txt
amqp==2.5.1
asn1crypto==1.0.1
billiard==3.5.0.5
celery==4.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorlog==3.1.4
Deprecated==1.2.7
idna==2.8
importlib-metadata==0.23
kombu==4.1.0
more-itertools==7.2.0
peewee==3.5.0
prettytable==0.7.2
pycparser==2.19
pycrypto==2.6.1
PyMySQL==0.8.1
pytz==2019.3
requests==2.22.0
redis==3.0.1
six==1.12.0
urllib3==1.22
vine==1.3.0
zipp==0.6.0
